requirements.txt
selenium~=4.1.0
discord.py==1.7.3